Use communicate() to ensure standard file descriptors are closed
This internally calls wait() but also ensures that any file descriptors
for stdin, stdout, and stderr are closed. If we don't do this we'll leak
file descriptors. For long running processes that start & stop a WebKit
server many times this can result in an eventual crash due to hitting
the max open files limit on the underlying system.